% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGG-hessian.R
\name{d2ldd2}
\alias{d2ldd2}
\title{Approximate second derivative with respect to sigma}
\usage{
d2ldd2(y, mu, sigma, nu)
}
\arguments{
\item{y}{Vector of observations.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{nu}{Shape parameter.}
}
\value{
Numeric vector.
}
\description{
Approximate second derivative with respect to sigma
}
