% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGG-hessian.R
\name{d2ldm2}
\alias{d2ldm2}
\title{Approximate second derivative with respect to mu}
\usage{
d2ldm2(y, mu, sigma, nu)
}
\arguments{
\item{y}{Vector of observations.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{nu}{Shape parameter.}
}
\value{
Numeric vector.
}
\description{
Computes the approximate second derivative of the LGG
log-density with respect to the location parameter.
}
