% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGG-hessian.R
\name{d2ldv2}
\alias{d2ldv2}
\title{Approximate second derivative with respect to nu}
\usage{
d2ldv2(y, mu, sigma, nu)
}
\arguments{
\item{y}{Vector of observations.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{nu}{Shape parameter.}
}
\value{
Numeric vector.
}
\description{
Approximate second derivative with respect to nu
}
