% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLGG.R
\name{dLGG}
\alias{dLGG}
\title{Log-Generalized Gamma Density Function}
\usage{
dLGG(y, mu = 0, sigma = 1, nu = 0.5, log = FALSE)
}
\arguments{
\item{y}{Vector of quantiles}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter (positive)}

\item{nu}{Shape parameter}

\item{log}{Logical; if TRUE, probabilities are given as log}
}
\value{
Density values
}
\description{
Log-Generalized Gamma Density Function
}
\examples{
dLGG(0, mu = 0, sigma = 1, nu = 0.5)
dLGG(0:2, mu = 0, sigma = 1, nu = 0.5, log = TRUE)
}
