% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssLGG-package.R
\docType{package}
\name{gamlssLGG-package}
\alias{gamlssLGG}
\alias{gamlssLGG-package}
\title{gamlssLGG: Log-Generalized Gamma Distribution for GAMLSS}
\description{
Implements the Log-Generalized Gamma (LGG) distribution for use with the GAMLSS framework. Provides density, distribution, quantile, random generation functions, analytical scores, and a GAMLSS family object.
}
\author{
\strong{Maintainer}: Rodolpho Jórdan Domingos Quintela \email{rodolpho.jordan@usp.br}

Authors:
\itemize{
  \item Rodolpho Jórdan Domingos Quintela \email{rodolpho.jordan@usp.br}
}

Other contributors:
\itemize{
  \item Gilberto Alvarenga Paula [contributor, thesis advisor]
}

}
\keyword{internal}
