% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGG.R
\name{LGG}
\alias{LGG}
\title{GAMLSS Family for the Log-Generalized Gamma Distribution}
\usage{
LGG(mu.link = "identity", sigma.link = "log", nu.link = "identity")
}
\arguments{
\item{mu.link}{Link function for mu}

\item{sigma.link}{Link function for sigma}

\item{nu.link}{Link function for nu}
}
\value{
A gamlss.family object
}
\description{
GAMLSS Family for the Log-Generalized Gamma Distribution
}
\examples{
\dontrun{
library(gamlss)
set.seed(123)
n <- 300
x <- runif(n)
mu <- 1 + 2*x
y <- rLGG(n, mu = mu, sigma = 0.5, nu = 0.4)
dados <- data.frame(y, x)
fit <- gamlss(y ~ x, family = LGG(), data = dados, trace = TRUE)
summary(fit)
}
}
