% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logC_LGG}
\alias{logC_LGG}
\title{Internal function for the log-normalizing constant}
\usage{
logC_LGG(nu)
}
\arguments{
\item{nu}{Shape parameter}
}
\description{
Internal function for the log-normalizing constant
}
\keyword{internal}
