% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLGG.R
\name{pLGG}
\alias{pLGG}
\title{Log-Generalized Gamma Distribution Function}
\usage{
pLGG(q, mu = 0, sigma = 1, nu = 0.5, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter (positive)}

\item{nu}{Shape parameter}

\item{lower.tail}{Logical; if TRUE, probabilities are P[X <= x]}

\item{log.p}{Logical; if TRUE, probabilities are given as log}
}
\value{
Cumulative distribution function values
}
\description{
Log-Generalized Gamma Distribution Function
}
\examples{
pLGG(0, mu = 0, sigma = 1, nu = 0.5)
pLGG(c(-1, 0, 1), mu = 0, sigma = 1, nu = 0.5)
}
