% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qLGG.R
\name{qLGG}
\alias{qLGG}
\title{Log-Generalized Gamma Quantile Function}
\usage{
qLGG(p, mu = 0, sigma = 1, nu = 0.5, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter (positive)}

\item{nu}{Shape parameter}

\item{lower.tail}{Logical; if TRUE, probabilities are P[X <= x]}

\item{log.p}{Logical; if TRUE, probabilities are given as log}
}
\value{
Quantile values
}
\description{
Log-Generalized Gamma Quantile Function
}
\examples{
qLGG(0.5, mu = 0, sigma = 1, nu = 0.5)
qLGG(c(0.25, 0.5, 0.75), mu = 0, sigma = 1, nu = 0.5)
}
