% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLGG.R
\name{rLGG}
\alias{rLGG}
\title{Random Generation for the Log-Generalized Gamma Distribution}
\usage{
rLGG(n, mu = 0, sigma = 1, nu = 0.5)
}
\arguments{
\item{n}{Number of observations}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter (positive)}

\item{nu}{Shape parameter}
}
\value{
Vector of random numbers
}
\description{
Random Generation for the Log-Generalized Gamma Distribution
}
\examples{
rLGG(10, mu = 0, sigma = 1, nu = 0.5)
}
