% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score_nu}
\alias{score_nu}
\title{Analytical Score for nu}
\usage{
score_nu(y, mu, sigma, nu)
}
\arguments{
\item{y}{Observations}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{nu}{Shape parameter}
}
\value{
Score values
}
\description{
Analytical Score for nu
}
