% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score_sigma}
\alias{score_sigma}
\title{Analytical Score for sigma}
\usage{
score_sigma(y, mu, sigma, nu)
}
\arguments{
\item{y}{Observations}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{nu}{Shape parameter}
}
\value{
Score values
}
\description{
Analytical Score for sigma
}
